+++ src/lib.rs
@@ -575,7 +575,7 @@ impl<K: 'static> OAuth2<K> {
     ///                 },
     ///                 "client id".to_string(),
     ///                 "client secret".to_string(),
-    ///                 "http://localhost:8000/auth".to_string(),
+    ///                 Some("http://localhost:8000/auth".to_string()),
     ///             );
     ///             Ok(rocket.attach(OAuth2::<MyProvider>::custom(HyperSyncRustlsAdapter, config)))
     ///         }))